moved settings to orchestratorlib
