+++ packages/orchestratorlib/src/index.ts
@@ -3,8 +3,8 @@
  * Licensed under the MIT License.
  */
 
+import {DataSourceHelper} from './datasourcehelper';
 import {LabelResolver} from './labelresolver';
-import {LuisQnaHelper} from './luisqnahelper';
 import {Orchestrator} from './orchestrator';
 import {OrchestratorAssess} from './assess';
 import {OrchestratorBaseModel} from './basemodel';
@@ -16,11 +16,12 @@ import {OrchestratorHelper} from './orchestratorhelper';
 import {OrchestratorPredict} from './predict';
 import {OrchestratorQuery} from './query';
 import {OrchestratorTest} from './test';
+import {OrchestratorDataSource, OrchestratorDataSourceSettings, OrchestratorSettings} from './settings';
 import {Utility} from './utility';
 
 export {
+  DataSourceHelper,
   LabelResolver,
-  LuisQnaHelper,
   Orchestrator,
   OrchestratorAssess,
   OrchestratorBaseModel,
@@ -30,6 +31,9 @@ export {
   // OrchestratorFineTune,
   OrchestratorPredict,
   OrchestratorQuery,
+  OrchestratorDataSource,
+  OrchestratorDataSourceSettings,
+  OrchestratorSettings,
   OrchestratorTest,
   OrchestratorHelper,
   Utility,
